LPS-51313 Refactor to do the logic in a subclass of CounterFinderImpl, so that we can selective wiring different impl on boot up, completely eliminate runtime overhead.
