+++ portal-impl/src/com/liferay/counter/service/persistence/impl/CounterFinderImpl.java
@@ -19,7 +19,6 @@ import com.liferay.counter.model.CounterHolder;
 import com.liferay.counter.model.CounterRegister;
 import com.liferay.counter.model.impl.CounterImpl;
 import com.liferay.counter.service.persistence.CounterFinder;
-import com.liferay.counter.service.util.MultiDataCenterCounterIncrementer;
 import com.liferay.portal.kernel.cache.CacheRegistryItem;
 import com.liferay.portal.kernel.cache.CacheRegistryUtil;
 import com.liferay.portal.kernel.concurrent.CompeteLatch;
@@ -60,15 +59,6 @@ public class CounterFinderImpl
 	@Override
 	public void afterPropertiesSet() {
 		CacheRegistryUtil.register(this);
-
-		if (PropsValues.COUNTER_DATA_CENTER_COUNT > 1) {
-			_multiDataCenterCounterIncrementer.initialize(
-				PropsValues.COUNTER_DATA_CENTER_COUNT,
-				PropsValues.COUNTER_DATA_CENTER_DEPLOYMENT_ID);
-		}
-		else {
-			_multiDataCenterCounterIncrementer = null;
-		}
 	}
 
 	@Override
@@ -123,15 +113,7 @@ public class CounterFinderImpl
 
 		CounterRegister counterRegister = getCounterRegister(name);
 
-		long value = _competeIncrement(counterRegister, size);
-
-		if (_multiDataCenterCounterIncrementer == null) {
-			return value;
-		}
-		else {
-			return _multiDataCenterCounterIncrementer.getMultiClusterSafeValue(
-				value);
-		}
+		return _competeIncrement(counterRegister, size);
 	}
 
 	@Override
@@ -213,12 +195,6 @@ public class CounterFinderImpl
 		CounterRegister counterRegister = createCounterRegister(name, size);
 
 		_counterRegisterMap.put(name, counterRegister);
-	}
-
-	public void setMultiDataCenterCounterIncrementer(
-		MultiDataCenterCounterIncrementer multiDataCenterCounterIncrementer) {
-
-		_multiDataCenterCounterIncrementer = multiDataCenterCounterIncrementer;
 	}
 
 	protected CounterRegister createCounterRegister(String name) {
@@ -452,8 +428,6 @@ public class CounterFinderImpl
 
 	private final Map<String, CounterRegister> _counterRegisterMap =
 		new ConcurrentHashMap<String, CounterRegister>();
-	private MultiDataCenterCounterIncrementer
-		_multiDataCenterCounterIncrementer;
 	private final Map<String, Integer> _rangeSizeMap =
 		new ConcurrentHashMap<String, Integer>();
